Creates the Cache class in system.process + this two dependencies : Attribute and Method
This class is used to register in a buffer a stack (LIFO) of attributes or methods to apply or invoke in a specified target object in the application. This class is a command who implement the Runnable interface to flush with the run() method all elements in the cache object.

// TODO finish the unit tests of the Cache class.